refactor: connect new service and delete old one
