--- src/main/java/nextstep/subway/maps/map/application/FareMapService.java
@@ -1,68 +1,0 @@
-package nextstep.subway.maps.map.application;
-
-import nextstep.subway.maps.line.application.LineService;
-import nextstep.subway.maps.line.domain.Line;
-import nextstep.subway.maps.line.dto.LineResponse;
-import nextstep.subway.maps.line.dto.LineStationResponse;
-import nextstep.subway.maps.map.domain.PathType;
-import nextstep.subway.maps.map.domain.SubwayPath;
-import nextstep.subway.maps.map.dto.FarePathResponse;
-import nextstep.subway.maps.map.dto.MapResponse;
-import nextstep.subway.maps.map.dto.PathResponseAssembler;
-import nextstep.subway.maps.station.application.StationService;
-import nextstep.subway.maps.station.domain.Station;
-import nextstep.subway.maps.station.dto.StationResponse;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
-
-@Service
-public class FareMapService {
-    private LineService lineService;
-    private StationService stationService;
-    private PathService pathService;
-    private FareCalculator fareCalculator;
-
-    public FareMapService(LineService lineService, StationService stationService, PathService pathService, FareCalculator fareCalculator) {
-        this.lineService = lineService;
-        this.stationService = stationService;
-        this.pathService = pathService;
-        this.fareCalculator = fareCalculator;
-    }
-
-    public MapResponse findMap() {
-        List<Line> lines = lineService.findLines();
-        Map<Long, Station> stations = findStations(lines);
-
-        List<LineResponse> lineResponses = lines.stream()
-                .map(it -> LineResponse.of(it, extractLineStationResponses(it, stations)))
-                .collect(Collectors.toList());
-
-        return new MapResponse(lineResponses);
-    }
-
-    public FarePathResponse findPath(Long source, Long target, PathType type) {
-        List<Line> lines = lineService.findLines();
-        SubwayPath subwayPath = pathService.findPath(lines, source, target, type);
-        Map<Long, Station> stations = stationService.findStationsByIds(subwayPath.extractStationId());
-
-        return PathResponseAssembler.assemble(subwayPath, stations, fareCalculator.calculate(subwayPath.calculateDistance()));
-    }
-
-    private Map<Long, Station> findStations(List<Line> lines) {
-        List<Long> stationIds = lines.stream()
-                .flatMap(it -> it.getStationInOrder().stream())
-                .map(it -> it.getStationId())
-                .collect(Collectors.toList());
-
-        return stationService.findStationsByIds(stationIds);
-    }
-
-    private List<LineStationResponse> extractLineStationResponses(Line line, Map<Long, Station> stations) {
-        return line.getStationInOrder().stream()
-                .map(it -> LineStationResponse.of(line.getId(), it, StationResponse.of(stations.get(it.getStationId()))))
-                .collect(Collectors.toList());
-    }
-}
