+++ src/main/java/nextstep/subway/maps/map/dto/PathResponseAssembler.java
@@ -9,23 +9,14 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 public class PathResponseAssembler {
-    public static PathResponse assemble(SubwayPath subwayPath, Map<Long, Station> stations) {
+
+    public static PathResponse assemble(SubwayPath subwayPath, Map<Long, Station> stations, int fare) {
         List<StationResponse> stationResponses = subwayPath.extractStationId().stream()
                 .map(it -> StationResponse.of(stations.get(it)))
                 .collect(Collectors.toList());
 
         int distance = subwayPath.calculateDistance();
 
-        return new PathResponse(stationResponses, subwayPath.calculateDuration(), distance);
-    }
-
-    public static FarePathResponse assemble(SubwayPath subwayPath, Map<Long, Station> stations, int fare) {
-        List<StationResponse> stationResponses = subwayPath.extractStationId().stream()
-                .map(it -> StationResponse.of(stations.get(it)))
-                .collect(Collectors.toList());
-
-        int distance = subwayPath.calculateDistance();
-
-        return new FarePathResponse(stationResponses, subwayPath.calculateDuration(), distance, fare);
+        return new PathResponse(stationResponses, subwayPath.calculateDuration(), distance, fare);
     }
 }
